backend_api/requirements.txt
fastapi
uvicorn[standard]
apscheduler
pymongo
boto3
mysql-connector-python
sqlmodel
psycopg2-binary
presidio-analyzer
presidio-anonymizer
presidio-structured
python-jose[cryptography]
passlib[argon2]
python-multipart
boto3
python-docx
pdf2image
pytesseract
pydantic-settings
python-dotenv
structlog
python-json-logger
spacy
